agent/requirements.txt
kubernetes==28.1.0
aiohttp==3.9.0
asyncio==4.0.0
PyYAML==6.0.1
